+++ src/main/java/com/mith/inn_orders_generator/dao/ComandaDaoImplements.java
@@ -4,5 +4,5 @@ import com.mith.inn_orders_generator.domain.Comanda;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public class ComandaDaoImplements extends AbstractDao<Comanda, Long> implements ComandaDao {
+public class ComandaDaoImplements extends AbstractDao<Comanda, Integer> implements ComandaDao {
 }
